# document/src/main/resources/swagger.yaml
---
swagger: "2.0"
info:
  description: "Competence centre project process microservice"
  version: "1.1.0"
  title: "document"
tags:
- name: "documents"
- name: "oproc"
- name: "processes"
- name: "state"
- name: "users"
paths:
  /documents:
    get:
      tags:
      - "documents"
      operationId: "getDocuments"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DocumentResource"
    post:
      tags:
      - "documents"
      operationId: "postDocument"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/DocumentResource"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DocumentResource"
  /documents/{id}:
    get:
      tags:
      - "documents"
      operationId: "getDocument"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DocumentResource"
    put:
      tags:
      - "documents"
      operationId: "putDocument"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/DocumentResource"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DocumentResource"
    delete:
      tags:
      - "documents"
      operationId: "deleteDocument"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DocumentResource"
    patch:
      tags:
      - "documents"
      operationId: "patchDocument"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/DocumentResource"
      - name: "userId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DocumentResource"
  /oproc:
    post:
      tags:
      - "oproc"
      operationId: "createObservableProcess"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ObservableProcessResource"
  /oproc/{id}:
    get:
      tags:
      - "oproc"
      operationId: "getObservableProcess"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ObservableProcessResource"
  /processes/{id}:
    get:
      tags:
      - "processes"
      operationId: "getProcess"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ProcessResource"
  /state/check:
    post:
      tags:
      - "state"
      operationId: "check"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/StateCheckRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/StateCheckResponse"
  /state/end:
    post:
      tags:
      - "state"
      operationId: "end"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/StateEndRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/StateEndResponse"
  /state/init:
    post:
      tags:
      - "state"
      operationId: "init"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/StateInitRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/StateInitResponse"
  /state/retrieveprice:
    post:
      tags:
      - "state"
      operationId: "retrievePrice"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/StateRetrievePriceRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/StateRetrievePriceResponse"
  /users:
    post:
      tags:
      - "users"
      operationId: "postUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/UserResource"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/UserResource"
  /users/{id}:
    get:
      tags:
      - "users"
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/UserResource"
definitions:
  DocumentResource:
    type: "object"
    properties:
      documentTitle:
        type: "string"
      signature:
        type: "string"
      links:
        type: "array"
        xml:
          name: "link"
          namespace: "http://www.w3.org/2005/Atom"
        readOnly: true
        items:
          xml:
            name: "link"
          $ref: "#/definitions/Link"
  Link:
    type: "object"
    properties:
      rel:
        type: "string"
        xml:
          attribute: true
      href:
        type: "string"
        xml:
          attribute: true
      hreflang:
        type: "string"
        xml:
          attribute: true
      media:
        type: "string"
        xml:
          attribute: true
      title:
        type: "string"
        xml:
          attribute: true
      type:
        type: "string"
        xml:
          attribute: true
      deprecation:
        type: "string"
        xml:
          attribute: true
  ObservableProcessResource:
    type: "object"
    properties:
      state:
        type: "string"
        enum:
        - "PUT_DOCUMENT_IN_PROGRESS"
        - "WAITED_IN_PROGRESS"
        - "END"
      links:
        type: "array"
        xml:
          name: "link"
          namespace: "http://www.w3.org/2005/Atom"
        readOnly: true
        items:
          xml:
            name: "link"
          $ref: "#/definitions/Link"
  ProcessResource:
    type: "object"
    properties:
      processCode:
        type: "string"
      processState:
        type: "string"
      links:
        type: "array"
        xml:
          name: "link"
          namespace: "http://www.w3.org/2005/Atom"
        readOnly: true
        items:
          xml:
            name: "link"
          $ref: "#/definitions/Link"
  StateCheckRequest:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
  StateCheckResponse:
    type: "object"
    properties:
      blocked:
        type: "boolean"
  StateEndRequest:
    type: "object"
  StateEndResponse:
    type: "object"
    properties:
      status:
        type: "string"
  StateInitRequest:
    type: "object"
  StateInitResponse:
    type: "object"
    properties:
      status:
        type: "string"
  StateRetrievePriceRequest:
    type: "object"
    properties:
      currency:
        type: "string"
  StateRetrievePriceResponse:
    type: "object"
    properties:
      currency:
        type: "string"
      rate:
        type: "string"
  UserResource:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
      links:
        type: "array"
        xml:
          name: "link"
          namespace: "http://www.w3.org/2005/Atom"
        readOnly: true
        items:
          xml:
            name: "link"
          $ref: "#/definitions/Link"